+++ product/server/common/src/main/java/io/codekvast/common/lock/impl/LockManagerImpl.java
@@ -21,21 +21,18 @@
  */
 package io.codekvast.common.lock.impl;
 
+import io.codekvast.common.lock.Lock;
 import io.codekvast.common.lock.LockManager;
 import io.codekvast.common.metrics.CommonMetricsService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.annotation.PostConstruct;
-import java.time.Duration;
-import java.time.Instant;
+import java.time.Clock;
 import java.util.Optional;
-import java.util.concurrent.ConcurrentHashMap;
 
 /**
  * @author olle.hallin@crisp.se
@@ -47,50 +44,31 @@ public class LockManagerImpl implements LockManager {
 
     private final JdbcTemplate jdbcTemplate;
     private final CommonMetricsService metricsService;
-
-    private final ConcurrentHashMap<Lock, Instant> locksAcquiredAt = new ConcurrentHashMap<>();
-
-    @PostConstruct
-    @Transactional
-    public void populateLocksTable() {
-        for (Lock lock : Lock.values()) {
-            int inserted = jdbcTemplate.update("INSERT IGNORE INTO internal_locks(name) VALUES(?)", lock.name());
-            if (inserted > 0) {
-                logger.info("Inserted {} to internal_locks table", lock.name());
-            }
-        }
-    }
+    private final Clock clock;
 
     @Override
     @Transactional(propagation = Propagation.MANDATORY)
-    public Optional<Lock> acquireLock(Lock lock) {
-        // TODO: Make locking more fine granular by using SELECT GET_LOCK() and SELECT RELEASE_LOCK
-        //   and include the customerId in the lock key.
-        try {
-            Instant acquireStartedAt = Instant.now();
-            String s =
-                jdbcTemplate.queryForObject("SELECT name FROM internal_locks WHERE name = ? FOR UPDATE WAIT ?", String.class, lock.name(), lock.getLockWaitSeconds());
-            logger.debug("Acquired lock {}", lock);
-            metricsService.recordLockWait(lock, Duration.between(acquireStartedAt, Instant.now()));
-            locksAcquiredAt.put(lock, acquireStartedAt);
-            return Optional.of(lock);
-        } catch (DataAccessException e) {
-            logger.info("Failed to acquire lock {} within {} s", lock, lock.getLockWaitSeconds());
-            metricsService.countLockFailure(lock);
-            return Optional.empty();
+    public Optional<Lock> acquireLock(final Lock lock) {
+        Integer locked = jdbcTemplate.queryForObject("SELECT GET_LOCK(?, ?)", Integer.class, lock.key(), lock.getMaxLockWaitSeconds());
+        if (locked.equals(1)) {
+            Lock result = lock.withAcquiredAt(clock.instant());
+            logger.debug("Acquired lock {}", result);
+            return Optional.of(result);
         }
+        logger.info("Failed to acquire lock {}", lock);
+        metricsService.countLockFailure(lock);
+        return Optional.empty();
     }
 
     @Override
     @Transactional(propagation = Propagation.MANDATORY)
     public void releaseLock(Lock lock) {
-        logger.debug("Releasing lock {}", lock);
-        Instant acquiredAt = locksAcquiredAt.remove(lock);
-        if (acquiredAt == null) {
-            logger.warn("Attempt to release lock {} which were not previously acquired", lock);
+        Integer unlocked = jdbcTemplate.queryForObject("SELECT RELEASE_LOCK(?)", Integer.class, lock.key());
+        if (unlocked.equals(1)) {
+            logger.debug("Released lock {}", lock);
+            metricsService.recordLockUsage(lock.withReleasedAt(clock.instant()));
         } else {
-            metricsService.recordLockDuration(lock, Duration.between(acquiredAt, Instant.now()));
+            logger.warn("Attempt to release lock {} which was not previously acquired", lock);
         }
-        // No-op against the database, the lock is automatically released when the transaction ends.
     }
 }