+++ product/server/dashboard/src/test/java/io/codekvast/dashboard/file_import/impl/PublicationImporterImplTest.java
@@ -1,7 +1,6 @@
 package io.codekvast.dashboard.file_import.impl;
 
 import io.codekvast.common.lock.LockManager;
-import io.codekvast.common.lock.LockTemplate;
 import io.codekvast.dashboard.file_import.CodeBaseImporter;
 import io.codekvast.dashboard.file_import.InvocationDataImporter;
 import io.codekvast.dashboard.file_import.PublicationImporter;
@@ -20,7 +19,6 @@ import javax.validation.Validator;
 import java.io.*;
 import java.net.URISyntaxException;
 import java.util.Collections;
-import java.util.Optional;
 
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
@@ -53,8 +51,7 @@ public class PublicationImporterImplTest {
     @Before
     public void beforeTest() {
         MockitoAnnotations.initMocks(this);
-        when(lockManager.acquireLock(LockManager.Lock.IMPORT)).thenReturn(Optional.of(LockManager.Lock.IMPORT));
-        this.publicationImporter = new PublicationImporterImpl(codeBaseImporter, invocationDataImporter, validator, metricsService, new LockTemplate(lockManager));
+        this.publicationImporter = new PublicationImporterImpl(codeBaseImporter, invocationDataImporter, validator, metricsService);
     }
 
     @Test
@@ -71,8 +68,6 @@ public class PublicationImporterImplTest {
 
         verify(codeBaseImporter).importPublication(any(CodeBasePublication3.class));
         verify(validator).validate(any());
-        verify(lockManager).acquireLock(LockManager.Lock.IMPORT);
-        verify(lockManager).releaseLock(LockManager.Lock.IMPORT);
         verifyNoMoreInteractions(codeBaseImporter, invocationDataImporter, validator);
     }
 
@@ -119,8 +114,6 @@ public class PublicationImporterImplTest {
         // then
         assertThat(handled, is(true));
         verify(invocationDataImporter).importPublication(any(InvocationDataPublication2.class));
-        verify(lockManager).acquireLock(LockManager.Lock.IMPORT);
-        verify(lockManager).releaseLock(LockManager.Lock.IMPORT);
         verify(validator).validate(any());
         verifyNoMoreInteractions(codeBaseImporter, invocationDataImporter, validator);
     }